DRY up tests. Create TestHelper.setPersonalDetails()
