+++ tests/actions/ReportTest.js
@@ -16,7 +16,6 @@ import * as TestHelper from '../utils/TestHelper';
 import Log from '../../src/libs/Log';
 import * as PersistedRequests from '../../src/libs/actions/PersistedRequests';
 import * as User from '../../src/libs/actions/User';
-import * as ReportUtils from '../../src/libs/ReportUtils';
 
 describe('actions/Report', () => {
     beforeAll(() => {
@@ -82,26 +81,7 @@ describe('actions/Report', () => {
                 User.subscribeToUserEvents();
                 return waitForPromisesToResolve();
             })
-            .then(() => {
-                const avatar = ReportUtils.getDefaultAvatar(TEST_USER_LOGIN);
-                const details = {
-                    accountID: TEST_USER_ACCOUNT_ID,
-                    login: TEST_USER_LOGIN,
-                    avatar,
-                    displayName: 'Test User',
-                    firstName: 'Test',
-                    lastName: 'User',
-                    pronouns: '',
-                    timezone: CONST.DEFAULT_TIME_ZONE,
-                    payPalMeAddress: '',
-                    phoneNumber: '',
-                    avatarHighResolution: avatar,
-                };
-                Onyx.merge(ONYXKEYS.PERSONAL_DETAILS, {
-                    [TEST_USER_LOGIN]: details,
-                });
-                return waitForPromisesToResolve();
-            })
+            .then(() => TestHelper.setPersonalDetails(TEST_USER_LOGIN, TEST_USER_ACCOUNT_ID))
             .then(() => {
                 // This is a fire and forget response, but once it completes we should be able to verify that we
                 // have an "optimistic" report action in Onyx.
@@ -196,26 +176,7 @@ describe('actions/Report', () => {
 
         // GIVEN a test user with initial data
         return TestHelper.signInWithTestUser(TEST_USER_ACCOUNT_ID, TEST_USER_LOGIN)
-            .then(() => {
-                const avatar = ReportUtils.getDefaultAvatar(TEST_USER_LOGIN);
-                const details = {
-                    accountID: TEST_USER_ACCOUNT_ID,
-                    login: TEST_USER_LOGIN,
-                    avatar,
-                    displayName: TEST_USER_LOGIN,
-                    firstName: 'Test',
-                    lastName: 'User',
-                    pronouns: '',
-                    timezone: CONST.DEFAULT_TIME_ZONE,
-                    payPalMeAddress: '',
-                    phoneNumber: '',
-                    avatarHighResolution: avatar,
-                };
-                Onyx.merge(ONYXKEYS.PERSONAL_DETAILS, {
-                    [TEST_USER_LOGIN]: details,
-                });
-                return waitForPromisesToResolve();
-            })
+            .then(() => TestHelper.setPersonalDetails(TEST_USER_LOGIN, TEST_USER_ACCOUNT_ID))
             .then(() => {
                 global.fetch = TestHelper.getGlobalFetchMock();
 
@@ -268,25 +229,7 @@ describe('actions/Report', () => {
                 User.subscribeToUserEvents();
                 return waitForPromisesToResolve();
             })
-            .then(() => {
-                const avatar = ReportUtils.getDefaultAvatar(USER_1_LOGIN);
-                const details = {
-                    accountID: USER_1_ACCOUNT_ID,
-                    login: USER_1_LOGIN,
-                    avatar,
-                    displayName: USER_1_LOGIN,
-                    firstName: 'Test',
-                    lastName: 'User',
-                    pronouns: '',
-                    timezone: CONST.DEFAULT_TIME_ZONE,
-                    payPalMeAddress: '',
-                    phoneNumber: '',
-                    avatarHighResolution: avatar,
-                };
-                Onyx.merge(ONYXKEYS.PERSONAL_DETAILS, {
-                    [USER_1_LOGIN]: details,
-                });
-            })
+            .then(() => TestHelper.setPersonalDetails(USER_1_LOGIN, USER_1_ACCOUNT_ID))
             .then(() => {
                 // When a Pusher event is handled for a new report comment
                 channel.emit(Pusher.TYPE.ONYX_API_UPDATE, [
