+++ tests/utils/TestHelper.js
@@ -1,6 +1,10 @@
+import Onyx from 'react-native-onyx';
+import CONST from '../../src/CONST';
 import * as Session from '../../src/libs/actions/Session';
 import HttpUtils from '../../src/libs/HttpUtils';
+import ONYXKEYS from '../../src/ONYXKEYS';
 import waitForPromisesToResolve from './waitForPromisesToResolve';
+import * as ReportUtils from '../../src/libs/ReportUtils';
 
 /**
  * Simulate signing in and make sure all API calls in this flow succeed. Every time we add
@@ -71,7 +75,34 @@ function getGlobalFetchMock() {
         });
 }
 
+/**
+ * @param {String} login
+ * @param {Number} accountID
+ * @returns {Promise}
+ */
+function setPersonalDetails(login, accountID) {
+    const avatar = ReportUtils.getDefaultAvatar(login);
+    const details = {
+        accountID,
+        login,
+        avatar,
+        displayName: 'Test User',
+        firstName: 'Test',
+        lastName: 'User',
+        pronouns: '',
+        timezone: CONST.DEFAULT_TIME_ZONE,
+        payPalMeAddress: '',
+        phoneNumber: '',
+        avatarHighResolution: avatar,
+    };
+    Onyx.merge(ONYXKEYS.PERSONAL_DETAILS, {
+        [login]: details,
+    });
+    return waitForPromisesToResolve();
+}
+
 export {
     getGlobalFetchMock,
     signInWithTestUser,
+    setPersonalDetails,
 };
